pyproject.toml
[tool.poetry]
name = "user-editor"
version = "0.1.0"
description = ""
authors = ["nikogamsahurdia <eleron96@gmail.com>"]
readme = "README.md"
packages = [{include = "user_editor"}]

[tool.poetry.dependencies]
python = "^3.11"
flask = "^2.3.2"
fastapi = "^0.100.1"
flask-login = "^0.6.2"
flask-sqlalchemy = "^3.0.5"
flask-migrate = "^4.0.4"
flask-wtf = "^1.1.1"
psycopg2 = "^2.9.7"
alembic = "^1.11.2"
bcrypt = "^4.0.1"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
